added silly mode button to index
can now toggle the mode off and on from the button in the main interface; go back and forth from normal mode to silly mode
