+++ app/_layout.tsx
@@ -1,5 +1,5 @@
-import { Stack } from "expo-router";
+import { Stack } from "expo-router"; // expo-router automatically handles the navigation
 
 export default function RootLayout() {
-  return <Stack />;
+  return <Stack />; // This will auto-handle routing
 }
