+++ app/index.tsx
@@ -1,126 +1,24 @@
-import React, { useState } from 'react';
-import { View, TextInput, Button, FlatList, Text, StyleSheet, Alert, Modal, TouchableOpacity } from 'react-native';
-import { Calendar } from 'react-native-calendars';
-import ReminderItem from './reminder_item';
-import { Reminder } from './reminder_data';
+import React from 'react';
+import { View, Text, TouchableOpacity, StyleSheet } from 'react-native';
+import { useRouter } from 'expo-router'; // Use expo-router for navigation
 
-const App = () => {
-  const [reminders, setReminders] = useState<Reminder[]>([]);
-  const [title, setTitle] = useState('');
-  const [description, setDescription] = useState('');
-  const [dueDate, setDueDate] = useState<Date | undefined>(undefined);
-  const [showCalendar, setShowCalendar] = useState(false);
-  const [isImportant, setIsImportant] = useState(false);
-  const [isSillyMode, setIsSillyMode] = useState(false); // new state for silly mode
+const HomeScreen = () => {
+  const router = useRouter(); // Get the router from expo-router
 
-  const handleAddReminder = () => {
-    if (!title || !description || !dueDate) {
-      Alert.alert('Error', 'Please fill in all fields.');
-      return;
-    }
-
-    const newReminder: Reminder = {
-      id: reminders.length + 1,
-      title,
-      description,
-      dueDate,
-      isImportant,
-    };
-
-    setReminders((prev) => [...prev, newReminder]);
-    setTitle('');
-    setDescription('');
-    setDueDate(undefined);
-    setIsImportant(false);
-  };
-
-  const onDayPress = (day: any) => {
-    const selectedDate = new Date(day.dateString);
-    setDueDate(selectedDate);
-    setShowCalendar(false);  // Close the calendar after selection
-  };
-
-  // Function to shuffle reminders for silly mode
-  const shuffleReminders = (reminders: Reminder[]) => {
-    const shuffled = [...reminders];
-    for (let i = shuffled.length - 1; i > 0; i--) {
-      const j = Math.floor(Math.random() * (i + 1));
-      [shuffled[i], shuffled[j]] = [shuffled[j], shuffled[i]]; // switches random elements into random places
-    }
-    return shuffled;
+  // Handle button press (navigating to Silly Mode Page)
+  const goToSillyMode = () => {
+    router.push('/sillymode_page'); // Navigate to Silly Mode page
   };
 
   return (
     <View style={styles.container}>
-      <Text style={styles.header}>SILLY REMINDERS!!!!!</Text>
-
-      {/* Reminder Input Form */}
-      <View style={styles.inputContainer}>
-        <TextInput
-          placeholder="ltiTe"
-          value={title}
-          onChangeText={setTitle}
-          style={styles.input}
-        />
-        <TextInput
-          placeholder="iispnterDoc"
-          value={description}
-          onChangeText={setDescription}
-          style={styles.input}
-        />
-        {/* Due Date */}
-        <TouchableOpacity style={styles.datePickerButton} onPress={() => setShowCalendar(true)}>
-          <Text style={styles.datePickerText}>
-            {dueDate ? dueDate.toLocaleDateString() : 'etDa Steuec leD'}
-          </Text>
-        </TouchableOpacity>
-
-        {/* Important Star */}
-        <TouchableOpacity onPress={() => setIsImportant(!isImportant)}>
-          <Text style={styles.importantText}>
-            {isImportant ? '⭐ Important' : '☆ r pM aasokrmnItat'}
-          </Text>
-        </TouchableOpacity>
-
-        <Button title="ADD A SILLY REMINDER?!" onPress={handleAddReminder} />
-      </View>
-
-      {/* Silly Mode Toggle */}
-      <TouchableOpacity
-        style={styles.sillyModeButton}
-        onPress={() => setIsSillyMode((prev) => !prev)}
-      >
-        <Text style={styles.sillyModeText}>
-          {isSillyMode ? 'Turn Off Silly Mode' : 'Turn On Silly Mode'}
-        </Text>
+      <Text style={styles.header}>Home Screen</Text>
+      <Text style={styles.sillyModeText}>Silly Mode is OFF</Text>
+      
+      {/* Circular button in top-right corner */}
+      <TouchableOpacity style={styles.sillyModeButton} onPress={goToSillyMode}>
+        <Text style={styles.buttonText}>OFF</Text>
       </TouchableOpacity>
-
-      {/* Modal for the Calendar Picker */}
-      <Modal visible={showCalendar} animationType="slide" transparent={true}>
-        <View style={styles.modalOverlay}>
-          <View style={styles.calendarContainer}>
-            <Calendar
-              onDayPress={onDayPress}
-              markedDates={{
-                [dueDate?.toISOString().split('T')[0] || '']: { selected: true, selectedColor: 'blue' },
-              }}
-              theme={{
-                selectedDayBackgroundColor: '#ab2b4d',
-                todayTextColor: '#ab2b4d',
-              }}
-            />
-            <Button title="CLOSE CALENDAR!" onPress={() => setShowCalendar(false)} />
-          </View>
-        </View>
-      </Modal>
-
-      {/* List of Reminders */}
-      <FlatList
-        data={isSillyMode ? shuffleReminders(reminders) : reminders} // Shuffle when silly mode is on
-        keyExtractor={(item) => item.id.toString()}
-        renderItem={({ item }) => <ReminderItem reminder={item} />}
-        contentContainerStyle={styles.list}
-      />
     </View>
   );
 };
@@ -128,72 +26,39 @@ const App = () => {
 const styles = StyleSheet.create({
   container: {
     flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
     padding: 20,
-    backgroundColor: '#ffd4e9',
+    backgroundColor: '#fff',
   },
   header: {
-    fontSize: 28,
+    fontSize: 24,
     fontWeight: 'bold',
-    color: 'purple',
     marginBottom: 20,
-    textAlign: 'center',
-  },
-  inputContainer: {
-    marginBottom: 20,
-  },
-  input: {
-    height: 40,
-    borderColor: '#ccc',
-    borderWidth: 1,
-    marginBottom: 10,
-    paddingHorizontal: 10,
-    borderRadius: 5,
-    backgroundColor: '#fff',
-  },
-  datePickerButton: {
-    height: 40,
-    borderColor: '#ccc',
-    borderWidth: 1,
-    marginBottom: 10,
-    justifyContent: 'center',
-    alignItems: 'center',
-    borderRadius: 5,
-    backgroundColor: '#fff',
-  },
-  datePickerText: {
-    color: '#666',
-  },
-  importantText: {
-    marginBottom: 10,
-    color: '#f39c12',
-  },
-  sillyModeButton: {
-    marginTop: 20,
-    padding: 10,
-    backgroundColor: '#eb4bd5',
-    borderRadius: 5,
-    alignItems: 'center',
   },
   sillyModeText: {
-    color: '#green',
     fontSize: 16,
+    marginBottom: 20,
   },
-  list: {
-    paddingBottom: 20,
-  },
-  modalOverlay: {
-    flex: 1,
-    justifyContent: 'center',
+  sillyModeButton: {
+    position: 'absolute', // Position the button in the top-right
+    top: 20,
+    right: 20,
+    width: 50, // Set the width and height for the circle
+    height: 50,
+    backgroundColor: 'rgba(255, 0, 0, 0.5)', // Semi-transparent red background
+    borderRadius: 25, // Circle shape
+    justifyContent: 'center', // Center the text inside
     alignItems: 'center',
-    backgroundColor: 'rgba(255, 91, 244, 0.5)',
+    shadowColor: '#000', // Optional shadow for button
+    shadowOffset: { width: 0, height: 2 },
+    shadowOpacity: 0.8,
+    shadowRadius: 4,
   },
-  calendarContainer: {
-    backgroundColor: 'red',
-    padding: 20,
-    borderRadius: 10,
-    width: 300,
-    alignItems: 'center',
+  buttonText: {
+    color: '#fff', // Text color inside the button
+    fontWeight: 'bold',
   },
 });
 
-export default App;
+export default HomeScreen;
