+++ package.json
@@ -19,6 +19,7 @@
     "@react-native-community/datetimepicker": "^8.2.0",
     "@react-navigation/bottom-tabs": "^7.2.0",
     "@react-navigation/native": "^7.0.14",
+    "@react-navigation/stack": "^7.1.1",
     "expo": "~52.0.25",
     "expo-blur": "~14.0.2",
     "expo-constants": "~17.0.4",
@@ -37,8 +38,10 @@
     "react-native-calendars": "^1.1308.0",
     "react-native-gesture-handler": "~2.20.2",
     "react-native-reanimated": "~3.16.1",
-    "react-native-safe-area-context": "4.12.0",
+    "react-native-router-flux": "^4.3.1",
+    "react-native-safe-area-context": "^4.14.1",
     "react-native-screens": "~4.4.0",
+    "react-native-svg": "^15.11.1",
     "react-native-vector-icons": "^10.2.0",
     "react-native-web": "~0.19.13",
     "react-native-webview": "13.12.5"
